refactor: update ChatBotPage to enhance job display and styling
- Increased max-width of BotMessageWrapper and TypingIndicator for improved layout.
- Changed InputContainer background to a light gray for better visibility.
- Removed the JobPopup component and replaced it with JobGrid for displaying job results.
- Updated message structure to include jobs directly, enhancing the user experience.
- Cleaned up unused state and related logic for a more streamlined component.
