+++ frontend/src/assets/bot-avatar.png
