+++ frontend/src/assets/sparky.png
