+++ frontend/src/components/ImageAssets.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import asuLogo from '../assets/images/asu-logo.png';
-import botAvatar from '../assets/images/bot-avatar.png';
-import userAvatar from '../assets/images/my_profile.png';
+import asuLogo from '../assets/asu-logo.png';
+import botAvatar from '../assets/bot-avatar.png';
+import userAvatar from '../assets/my_profile.png';
 
 export const ASULogoImage: React.FC<{ className?: string }> = ({ className }) => (
   <img 
@@ -43,21 +43,3 @@ export const BotAvatarImage: React.FC<{ className?: string }> = ({ className }) 
     }} 
   />
 );
-
-// Instructions for adding custom images:
-/*
-1. Add your image files to src/assets/images/
-   - asu-logo.png (for ASU logo)
-   - user-avatar.jpg (for user profile picture)
-   - bot-avatar.png (for chatbot avatar)
-
-2. Import them at the top of this file:
-   import asuLogo from '../assets/images/asu-logo.png';
-
-3. Replace the placeholder components with img tags:
-   export const ASULogoImage: React.FC<{ className?: string }> = ({ className }) => (
-     <img src={asuLogo} alt="ASU Logo" className={className} />
-   );
-
-4. Update the ChatBotPage.tsx to use these components instead of the current placeholders
-*/
